Move reporting unit endpoints to own resource folder
