+++ ras_backstage/__init__.py
@@ -25,7 +25,7 @@ collection_exercise_api = Namespace('collection-exercise',
 collection_instrument_api = Namespace('collection-instrument',
                                       path='/backstage-api/v1/collection-instrument')
 party_api = Namespace('party', path='/backstage-api/v1/party')
-reporting_unit_api = Namespace('party', path='/backstage-api/v1/reporting-unit')
+reporting_unit_api = Namespace('reporting-unit', path='/backstage-api/v1/reporting-unit')
 sample_api = Namespace('sample', path='/backstage-api/v1/sample')
 secure_messaging_api = Namespace('secure-messaging',
                                  path='/backstage-api/v1/secure-message')
@@ -49,7 +49,7 @@ from ras_backstage.resources.collection_exercise.single_collection_exercise impo
 from ras_backstage.resources.collection_instrument.collection_instrument import CollectionInstrument  # NOQA # pylint: disable=wrong-import-position
 from ras_backstage.resources.info import Info  # NOQA # pylint: disable=wrong-import-position
 from ras_backstage.resources.party.get_party_details import PartyDetails  # NOQA # pylint: disable=wrong-import-position
-from ras_backstage.resources.party.get_reporting_unit import GetReportingUnit  # NOQA # pylint: disable=wrong-import-position
+from ras_backstage.resources.reporting_unit.get_reporting_unit import GetReportingUnit  # NOQA # pylint: disable=wrong-import-position
 from ras_backstage.resources.sample.sample import Sample  # NOQA # pylint: disable=wrong-import-position
 from ras_backstage.resources.secure_messaging.get_message_list import GetMessagesList  # NOQA # pylint: disable=wrong-import-position
 from ras_backstage.resources.secure_messaging.get_message import GetMessage  # NOQA # pylint: disable=wrong-import-position
